ggml-cpu: move s390x typedef to own header file
